split file components
