+++ app/react/V2/Routes/Entity/Components/relationships/rows/CollapsibleRelationshipRow.tsx
@@ -1,25 +1,22 @@
-import React, { Children, useState, type ReactNode } from 'react';
+import React, { useState, type ReactNode } from 'react';
 import { ChevronDownIcon, LinkIcon } from '@heroicons/react/24/outline';
 import { Translate, t } from '#app/I18N/index.js';
 import { ListCardRow } from '#V2/Components/UI/ListCardRow.js';
-import { TemplatePill } from '#V2/Components/UI/TemplatePill.js';
-import type {
-  RelationshipsPanelView,
-  RelationshipsPanelZoom,
-} from '#V2/Routes/Entity/Components/context/index.js';
+import type { RelationshipsPanelZoom } from '#V2/Routes/Entity/Components/context/index.js';
 import { useRelationshipsPanelLayout } from '#V2/Routes/Entity/Components/context/index.js';
 import { RelationshipRowCheckbox } from './RelationshipRowCheckbox.js';
-import { DirectionGlyph } from './DirectionGlyph.js';
 import { useRelationshipRowVisibility } from '../hooks/useRelationshipRowVisibility.js';
-
-type Direction = 'incoming' | 'outgoing' | 'both';
+import {
+  resolveCollapsibleRowHeader,
+  type HeaderContext,
+} from './CollapsibleRelationshipRowHeader.js';
 
 type CollapsibleRelationshipRowProps = {
   checkboxId: string;
   evidenceCount: number;
   header: ReactNode;
   meta?: ReactNode;
-  glyphDirection?: Direction;
+  glyphDirection?: HeaderContext['glyphDirection'];
   relationshipTypeName?: string;
   headerWrap?: boolean;
   isHub?: boolean;
@@ -27,118 +24,8 @@ type CollapsibleRelationshipRowProps = {
   targetTemplateId?: string;
   entityTitle?: string;
   templateName?: string;
+  onHeaderClick?: () => void;
   children: ReactNode;
-};
-
-type HeaderContext = {
-  hideTargetPill: boolean;
-  hideTemplateName: boolean;
-  view: RelationshipsPanelView;
-  zoom: RelationshipsPanelZoom;
-  header: ReactNode;
-  targetTemplateId?: string;
-  entityTitle?: string;
-  templateName?: string;
-  glyphDirection?: Direction;
-  relationshipTypeName?: string;
-  isHub: boolean;
-  memberCount: number;
-  headerWrap: boolean;
-};
-
-const entityTitleClass: Record<RelationshipsPanelZoom, string> = {
-  overview: 'text-xs',
-  compact: 'text-xs',
-  detail: 'text-sm',
-};
-
-const resolveAggregateHeader = (ctx: HeaderContext): ReactNode => {
-  if (!ctx.entityTitle || !ctx.targetTemplateId) return ctx.header;
-  if (ctx.view === 'tree' && ctx.zoom !== 'overview') {
-    return (
-      <>
-        {!ctx.hideTemplateName && (
-          <TemplatePill templateId={ctx.targetTemplateId} label={ctx.templateName} />
-        )}
-        <span
-          title={ctx.entityTitle}
-          className={`min-w-0 truncate font-medium text-ink ${entityTitleClass[ctx.zoom]}`}
-        >
-          {ctx.entityTitle}
-        </span>
-      </>
-    );
-  }
-  return <TemplatePill templateId={ctx.targetTemplateId} label={ctx.entityTitle} />;
-};
-
-const resolveHeader = (ctx: HeaderContext): ReactNode => {
-  if (ctx.entityTitle && ctx.targetTemplateId) return resolveAggregateHeader(ctx);
-  return ctx.header;
-};
-
-const fallbackLabel = ({
-  glyphDirection,
-  relationshipTypeName,
-}: Pick<HeaderContext, 'glyphDirection' | 'relationshipTypeName'>) => {
-  const relLabel = relationshipTypeName ? (
-    <span className="truncate text-[10px] capitalize text-ink-tertiary">
-      {relationshipTypeName}
-    </span>
-  ) : null;
-
-  if (!glyphDirection) return relLabel;
-
-  return (
-    <span className="flex min-w-0 items-center gap-1.5 text-xs capitalize text-ink-secondary">
-      <DirectionGlyph direction={glyphDirection} />
-      {relLabel}
-    </span>
-  );
-};
-
-const overviewHeader = (ctx: HeaderContext) => {
-  if (ctx.hideTargetPill && !ctx.isHub) return fallbackLabel(ctx);
-  if (ctx.isHub && ctx.memberCount > 3) {
-    const pills = Children.toArray(ctx.header).slice(0, 3);
-    return (
-      <>
-        {pills}
-        <span className="text-[10px] text-ink-tertiary">+{ctx.memberCount - 3}</span>
-      </>
-    );
-  }
-  return resolveHeader(ctx);
-};
-
-const compactHeader = (ctx: HeaderContext) => {
-  if (ctx.hideTargetPill && !ctx.isHub) return fallbackLabel(ctx);
-  if (ctx.isHub && ctx.memberCount > 3) {
-    const pills = Children.toArray(ctx.header).slice(0, 3);
-    return (
-      <>
-        {pills}
-        <span className="text-[10px] text-ink-tertiary">+{ctx.memberCount - 3}</span>
-      </>
-    );
-  }
-  const relLabel = ctx.relationshipTypeName ? (
-    <span className="truncate text-[10px] capitalize text-ink-tertiary">
-      {ctx.relationshipTypeName}
-    </span>
-  ) : null;
-  return (
-    <>
-      {ctx.glyphDirection && <DirectionGlyph direction={ctx.glyphDirection} />}
-      {resolveHeader(ctx)}
-      {relLabel}
-    </>
-  );
-};
-
-const detailHeader = (ctx: HeaderContext) => {
-  if (ctx.hideTargetPill) return fallbackLabel(ctx);
-  return resolveHeader(ctx);
 };
 
 const rowPadding: Record<RelationshipsPanelZoom, string> = {
@@ -160,13 +47,21 @@ const CollapsibleRelationshipRow = ({
   targetTemplateId,
   entityTitle,
   templateName,
+  onHeaderClick,
   children,
 }: CollapsibleRelationshipRowProps) => {
   const [expanded, setExpanded] = useState(false);
   const { view, zoom } = useRelationshipsPanelLayout();
   const { hideTargetPill, hideTemplateName } = useRelationshipRowVisibility();
   const toggle = () => setExpanded(current => !current);
-  const headerCtx: HeaderContext = {
+  const handleHeaderClick = () => {
+    if (onHeaderClick) {
+      onHeaderClick();
+      return;
+    }
+    toggle();
+  };
+  const headerContent = resolveCollapsibleRowHeader({
     hideTargetPill,
     hideTemplateName,
     view,
@@ -180,15 +75,7 @@ const CollapsibleRelationshipRow = ({
     isHub,
     memberCount,
     headerWrap,
-  };
-  let headerContent: ReactNode;
-  if (zoom === 'overview') {
-    headerContent = overviewHeader(headerCtx);
-  } else if (zoom === 'compact') {
-    headerContent = compactHeader(headerCtx);
-  } else {
-    headerContent = detailHeader(headerCtx);
-  }
+  });
   const chevron = (
     <ChevronDownIcon
       className={`h-3 w-3 shrink-0 text-ink-muted transition-transform ${
@@ -222,7 +109,7 @@ const CollapsibleRelationshipRow = ({
 
   return (
     <div className="border-b border-border/50 last:border-b-0">
-      <ListCardRow selected={false} onClick={toggle} className={rowPadding[zoom]}>
+      <ListCardRow selected={false} onClick={handleHeaderClick} className={rowPadding[zoom]}>
         <div className={`flex ${alignItems} justify-between gap-2`}>
           <div className={`flex min-w-0 items-center gap-1.5 ${headerWrap ? 'flex-wrap' : ''}`}>
             <RelationshipRowCheckbox relationshipId={checkboxId} />
